+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [0.1.5-dev][Unreleased] - Unreleased
+* Use `src-git-full` instead of `src-git` in feeds.conf
 
 ## [0.1.4] - 2022-05-30
 * Update to v1.0.7 of the packages-cci Docker image (includes public