+++ Changelog.md
@@ -9,6 +9,7 @@
 * Always add `packages` as a custom feed
 * Add `USE_GITHUB_FEEDS` run-time option
 * Add script to set ownership of bind mounts
+* Fix config options not set correctly
 
 ## [0.1.4] - 2022-05-30
 * Update to v1.0.7 of the packages-cci Docker image (includes public